+++ mrs/allocation/auctioneer.py
@@ -21,7 +21,7 @@ specified in the config file
 
 class Auctioneer(SimulatorInterface):
 
-    def __init__(self, timetable_manager, closure_window=5, **kwargs):
+    def __init__(self, timetables, closure_window=5, **kwargs):
         simulator = kwargs.get('simulator')
         super().__init__(simulator)
 
@@ -29,7 +29,7 @@ class Auctioneer(SimulatorInterface):
         self.api = kwargs.get('api')
         self.ccu_store = kwargs.get('ccu_store')
         self.robot_ids = list()
-        self.timetable_manager = timetable_manager
+        self.timetables = timetables
 
         self.closure_window = timedelta(minutes=closure_window)
         self.alternative_timeslots = kwargs.get('alternative_timeslots', False)
@@ -55,10 +55,10 @@ class Auctioneer(SimulatorInterface):
     def register_robot(self, robot_id):
         self.logger.debug("Registering robot %s", robot_id)
         self.robot_ids.append(robot_id)
-        self.timetable_manager.register_robot(robot_id)
+        self.timetables.register_robot(robot_id)
 
     def set_ztp(self, time_):
-        self.timetable_manager.ztp = time_
+        self.timetables.ztp = time_
 
     def run(self):
         if self.tasks_to_allocate and self.round.finished:
@@ -107,7 +107,7 @@ class Auctioneer(SimulatorInterface):
             task = self.tasks_to_allocate.pop(self.winning_bid.task_id)
             self.allocated_tasks[task.task_id] = task
 
-            self.timetable_manager.update_timetable(self.winning_bid.robot_id,
+            self.timetables.update_timetable(self.winning_bid.robot_id,
                                                     self.winning_bid.get_allocation_info(),
                                                     task)
 
@@ -182,7 +182,7 @@ class Auctioneer(SimulatorInterface):
 
         earliest_admissible_time = TimeStamp()
         earliest_admissible_time.timestamp = self.get_current_time() + timedelta(minutes=1)
-        task_announcement = TaskAnnouncement(tasks, self.round.id, self.timetable_manager.ztp, earliest_admissible_time)
+        task_announcement = TaskAnnouncement(tasks, self.round.id, self.timetables.ztp, earliest_admissible_time)
 
         self.logger.debug("Starting round: %s", self.round.id)
         self.logger.debug("Number of tasks to allocate: %s", len(tasks))
@@ -249,21 +249,21 @@ class Auctioneer(SimulatorInterface):
         # For now, returning the start navigation time from the dispatchable graph
         task_schedule = dict()
 
-        timetable = self.timetable_manager.timetables.get(robot_id)
+        timetable = self.timetables.timetables.get(robot_id)
 
         relative_start_time = timetable.dispatchable_graph.get_time(task_id, "start")
         relative_pickup_time = timetable.dispatchable_graph.get_time(task_id, "pickup")
         relative_latest_delivery_time = timetable.dispatchable_graph.get_time(task_id, "delivery", False)
 
         self.logger.debug("Current time %s: ", TimeStamp())
-        self.logger.debug("ztp %s: ", self.timetable_manager.ztp)
+        self.logger.debug("ztp %s: ", self.timetables.ztp)
         self.logger.debug("Relative start navigation time: %s", relative_start_time)
         self.logger.debug("Relative pickup time: %s", relative_pickup_time)
         self.logger.debug("Relative latest delivery time: %s", relative_latest_delivery_time)
 
-        start_time = self.timetable_manager.ztp + timedelta(seconds=relative_start_time)
-        pickup_time = self.timetable_manager.ztp + timedelta(seconds=relative_pickup_time)
-        delivery_time = self.timetable_manager.ztp + timedelta(seconds=relative_latest_delivery_time)
+        start_time = self.timetables.ztp + timedelta(seconds=relative_start_time)
+        pickup_time = self.timetables.ztp + timedelta(seconds=relative_pickup_time)
+        delivery_time = self.timetables.ztp + timedelta(seconds=relative_latest_delivery_time)
 
         self.logger.debug("Task %s start time: %s", task_id, start_time)
         self.logger.debug("Task %s pickup time : %s", task_id, pickup_time)
