+++ mrs/config/builder.py
@@ -9,8 +9,7 @@ from mrs.execution.fleet_monitor import FleetMonitor
 from mrs.execution.schedule_monitor import ScheduleMonitor
 from mrs.performance.tracker import PerformanceTracker
 from mrs.simulation.simulator import Simulator
-from mrs.timetable.timetable import Timetable
-from mrs.timetable.manager import TimetableManager
+from mrs.timetable.timetable import Timetable, Timetables
 from mrs.timetable.monitor import TimetableMonitor
 from planner.planner import Planner
 from stn.stp import STP
@@ -20,7 +19,7 @@ class MRTABuilder:
 
     _component_modules = {'simulator': Simulator,
                           'timetable': Timetable,
-                          'timetable_manager': TimetableManager,
+                          'timetables': Timetables,
                           'planner': Planner,
                           'delay_recovery': DelayRecovery,
                           'auctioneer': Auctioneer,
@@ -35,7 +34,7 @@ class MRTABuilder:
 
     _config_order = ['simulator',
                      'timetable',
-                     'timetable_manager',
+                     'timetables',
                      'planner',
                      'delay_recovery',
                      'auctioneer',
